ui tests: update old tests
